# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Integrity-First - OpenAPI 3.0",
    "description": "This is a Intergity-First Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "taha.muhammad.official@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    },
    {
      "url": "https://staging.integrity1auto.com"
    },
    {
      "url": "https://api.integrity1auto.com"
    }
  ],
...
  "paths": {
    "/auth/send-phonecode": {
      "post": {
        "tags": ["Register"],
        "summary": "Send code to provided phone number for verification",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phoneNumber", "shopId", "password"],
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+923204858850"
                  },
                  "password": {
                    "type": "string",
                    "example": "17283687126783"
                  },
                  "shopId": {
                    "type": "string",
                    "example": "63cb0a356e797714355c76c6"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/auth/verify-phonecode": {
      "post": {
        "tags": ["Register"],
        "summary": "Verify phone code",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phoneNumber", "code"],
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+923204858850"
                  },
                  "code": {
                    "type": "string",
                    "example": "123456"
                  },
                  "tokenData": {
                    "type": "object",
                    "properties": {
                      "deviceId": {
                        "type": "string",
                        "example": "12378127731982739818sa89798sd7f98s7d89f"
                      },
                      "fcmToken": {
                        "type": "string",
                        "example": "12378127731982739818sa89798sd7f98s7d89f12378127731982739818sa89798sd7f98s7d89f"
                      },
                      "osName": {
                        "type": "string",
                        "example": "IOS | ANDROID"
                      }
                    }
                  },
                  "deviceType": {
                    "type": "string",
                    "example": "IOS | ANDROID"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": ["Login"],
        "summary": "Login to the app",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phoneNumber", "password"],
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+923204858850"
                  },
                  "tokenData": {
                    "type": "object",
                    "properties": {
                      "deviceId": {
                        "type": "string",
                        "example": "12378127731982739818sa89798sd7f98s7d89f"
                      },
                      "fcmToken": {
                        "type": "string",
                        "example": "12378127731982739818sa89798sd7f98s7d89f12378127731982739818sa89798sd7f98s7d89f"
                      },
                      "osName": {
                        "type": "string",
                        "example": "IOS | ANDROID"
                      }
                    }
                  },
                  "password": {
                    "type": "string",
                    "example": "12sdkljfisd78979cx8v"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/auth/send-forgot-phonecode": {
      "post": {
        "tags": ["Forgot Password"],
        "summary": "Send code to provided phone number for verification",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phoneNumber"],
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+923204858850"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/auth/verify-forgot-phonecode": {
      "post": {
        "tags": ["Forgot Password"],
        "summary": "Verify forgot-password phone-code",
        "description": "In response you will get the token, which you can use to call update profile API with password in body to change passowrd",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phoneNumber", "code"],
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+923204858850"
                  },
                  "code": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/auth": {
      "delete": {
        "tags": ["Profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Call this API to delete logged-in user profile",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": ["Change Password"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Call this API to change password",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["oldPassword", "newPassword"],
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "example": "My old password"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "My new password"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/blog": {
      "get": {
        "tags": ["Blogs"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "summary": "Get the list all blogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": ["Blogs"],
        "summary": "Create a new blog",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Some Title"
                  },
                  "imageURL": {
                    "type": "string",
                    "example": "Some URL here"
                  },
                  "ctaKind": {
                    "type": "string",
                    "example": "Some kind"
                  },
                  "ctaValue": {
                    "type": "string",
                    "example": "Some CTA value"
                  },
                  "description": {
                    "type": "string",
                    "example": "Some description"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/blog/{blogId}": {
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": ["Blogs"],
        "summary": "Update blog",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/blog/{blogIds}": {
      "delete": {
        "tags": ["Blogs"],
        "summary": "Delete multiple blogs",
        "parameters": [
          {
            "name": "blogIds",
            "in": "path",
            "schema": {
              "type": "string",
              "example": "64065fe7f8db2e031d80fcc5,64065ffef8db2e031d80fd60"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "WebhookAPIKeyAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}